pyproject.toml
[project]
name = "bunq-hackathon-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.97.0",
    "bunq-sdk>=1.28.0",
    "django>=6.0.4",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "dotenv>=0.9.9",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.12",
    "requests>=2.33.1",
]
